Extract common functions to a separate module
